#159: Various UX improvements
Stopped using the command stack for in-zoom commands, because it is not
part of what is done in the current editor. This improves usability
because before this change when a thing was in-zoomed the editor became
dirty and had to be saved. Now only the new OPD must be saved.

Task-Url:
http://github.com/vainolo/Object-Process-Programming/issues/issue/159