Fixed a nasty bug in the process execution. Processes were not executed in the correct order because the dependencies were not transitive... After searching for many ways to fix this, decided that the most simple way is to modify how processes are executed, making this a lot simpler.
Basically, we now have a program counter that goes from top to bottom
executing processes as they come and not moving under processes that are
waiting or finished. If there are events, the execution is stopped, the
invoked process is run and then execution continues from below the
executed process.

I'll have to rewrite half my thesis :-(